fix: 在删除构建目录前检查其是否存在
避免在构建目录不存在时脚本报错

+++ build_run_android.ps1
@@ -3,7 +3,9 @@ Set-StrictMode -Version Latest
 $ErrorActionPreference = "Stop"
 
 # 清空构建目录
-Remove-Item -Path .\build -Recurse
+if (Test-Path -Path .\build) {
+    Remove-Item -Path .\build -Recurse
+}
 
 # Go 构建安卓可执行文件
 $env:GOOS = "linux"
